feat(elreal): high-precision euler_gamma via Brent-McMillan (#1053) (#1073)
* feat(elreal): high-precision euler_gamma via Brent-McMillan (#1053)

euler_gamma_zbcl was a stub returning a bare double literal (~17 digits)
while every other constant reached ~280; the #1048 exact-dyadic oracle
measured exactly that. Replace it with a real Brent-McMillan B1 generator:

  gamma = A(n)/B(n) - ln(n),
  B(n) = sum_k (n^k/k!)^2,  A(n) = sum_k (n^k/k!)^2 * H_k,  H_0 = 0,
  |error| ~ pi*exp(-4n)  ->  n >= D*ln(10)/4 for D digits.

Validated correct against s_euler_gamma: 47/64/129/194 digits at depth
3/4/8/12 (~16 digits/depth), reaching the host-double ceiling (~320, >= the
test's 300-digit threshold) at depth 20.

This is an eager-batch generator like the other constants. It is the
slowest of them (the Euler-Mascheroni constant has no rapidly-converging
elementary series, and ln(n) goes through the shared eager ln2/odd_power_
series). Measured bottlenecks at depth 12: the 560-term mul(w_k,H_k) loop
(132s) and ln2_zbcl (107s) -- both eager, both shared. Migrating the series
layer to the online (pull-driven) mul/div is #1061 Phase 3 and will speed up
euler_gamma AND pi/ln2/ln10 uniformly; doing it for euler_gamma alone would
leave it ln2-bound and create a one-off. So this ships the correct generator
now (the actual gap was correctness, not speed) and defers the uniform
speedup to #1061. The test is REGRESSION_LEVEL_4 (stress) only.

Also fixes a prerequisite latent compile bug: elreal_reference_digits.hpp
did `d.scale += b.exp` (int += integer<256>) -- broken since #1066 widened
the block exponent. It compiled only because default builds are LEVEL_1,
which #if-compiles-out the agreed_decimal_digits calls; at LEVEL_4 the
high-precision test did not compile at all. Cast b.exp to int (the per-block
offset fits int for any real host).

Re-enable euler_gamma in constants_highprecision.cpp. LEVEL_4 path compiles
on gcc and clang.

Resolves #1053


* fix(elreal): scope euler_gamma host check to in-range hosts/depth (#1053)

The basic el_math_constants test (LEVEL_1, fast CI) was exercising the new
euler_gamma_zbcl at kConstDepth=6 on double/float/bfloat16, which regressed
fast CI two ways:

  1. Correctness: Brent-McMillan's intermediate w_k = (n^k/k!)^2 peaks at
     ~exp(2n); the block significand carries that magnitude, so a narrow
     exponent host overflows once n exceeds ~44. float overflowed at depth 6
     (const<float> egamma = -0.868536), and bfloat16's 7-bit significand
     cannot carry the H_k/A/B/ln(n) computation at all (value oscillates ~5%
     around 0.5772 with depth -- any pass is coincidental).
  2. Cost: euler_gamma is O(n) eager terms; at depth 6 it added ~30-60s to a
     fast-tier test.

Fix is test-side (the double generator itself is unchanged and still validated
to 307 digits in the LEVEL_4 test): the host-tolerance smoke check now runs
euler_gamma only on hosts with float precision or better (digits >= 24, so
double + float, skip bfloat16) at a shallow depth 2 -- fast and within float's
exponent range (n=40 < 44), clearing the 1e-6/1e-12 tolerances (~14 digits on
float). Document the narrow-host range limitation in euler_gamma_zbcl.

el_math_constants PASS on gcc and clang.

Relates to #1053


---------

+++ include/sw/universal/verification/elreal_reference_digits.hpp
@@ -46,7 +46,10 @@ inline dyadic zbcl_to_dyadic(const ZBCL<FpType>& z, std::size_t maxBlocks = 512)
 	dyadic acc;  // 0
 	for (const auto& b : z.take(maxBlocks)) {
 		dyadic d = dyadic::from_double(static_cast<double>(b.v));
-		d.scale += b.exp;       // block value == v * 2^exp
+		d.scale += static_cast<int>(b.exp);   // block value == v * 2^exp; b.exp is
+		                                      // a wide integer<256> (#1066) -- the
+		                                      // offset fits int for any real host
+		                                      // (combined exponent ~ +/- 1074).
 		acc = acc + d;
 	}
 	return acc;
